+++ CHANGELOG.md
@@ -20,6 +20,7 @@
   `InputMediaPhoto`, `InputMediaVideo` types.
 - New #32: Add `setMessageReaction` method and `ReactionType::toRequestArray()` method.
 - New #33: Add `banChatMember` and `unbanChatMember` methods.
+- New #34: Add `restrictChatMember` method and `ChatPermissions::toRequestArray()` method.
 - Chg #24: Move update methods to `Vjik\TelegramBot\Api\Method\Update` namespace, and update types to
   `Vjik\TelegramBot\Api\Type\Update` namespace.
 - Chg #30: Remove `TelegramRequestWithFilesInterface`.